services/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-dotenv==1.0.0
PyYAML==6.0.1
httpx==0.25.2
PyJWT==2.8.0
aioredis==2.0.1
starlette==0.27.0
prometheus-client==0.19.0
psutil==5.9.6
